FIX: Reorganize
