+++ mne/externals/h5io/__init__.py
@@ -1,0 +1,6 @@
+"""Python Objects Onto HDF5
+"""
+
+__version__ = '0.1.dev0'
+
+from ._h5io import read_hdf5, write_hdf5, _TempDir, object_diff  # noqa, analysis:ignore
